refactor(native): group merchant runtime by subdomain
