+++ src/native/engine/engine.c
@@ -22,8 +22,8 @@
 #include "runtime/player_motion_host.h"
 #include "runtime/player_api.h"
 #include "runtime/player_build_abi.h"
-#include "runtime/merchant_query_abi.h"
-#include "runtime/merchant_trade_domain.h"
+#include "runtime/merchant/merchant_query_abi.h"
+#include "runtime/merchant/merchant_trade_domain.h"
 #include "runtime/player_resource_abi.h"
 #include "runtime/player_merchant_adapter.h"
 #include "runtime/player_sync_abi.h"
