+++ src/native/engine/runtime/merchant/merchant_trade_domain.c
