+++ src/native/engine/runtime/player_gateway_abi.h
@@ -1,7 +1,7 @@
 #ifndef BANANA_ENGINE_RUNTIME_PLAYER_GATEWAY_ABI_H
 #define BANANA_ENGINE_RUNTIME_PLAYER_GATEWAY_ABI_H
 
-#include "merchant_service.h"
+#include "merchant/merchant_service.h"
 #include "player_registry.h"
 
 #ifdef __cplusplus
