+++ tests/native/runtime/merchant/runtime_merchant_trade_abi_test.c
@@ -1,4 +1,4 @@
-#include "runtime/merchant_trade_abi.h"
+#include "runtime/merchant/merchant_trade_abi.h"
 
 #include <stdio.h>
