+++ tests/native/runtime_compile_guard_abi_test.c
@@ -4,8 +4,8 @@
 #include "runtime/player_gateway_abi.h"
 #include "runtime/player_build_abi.h"
 #include "runtime/player_builds.h"
-#include "runtime/merchant_query_abi.h"
-#include "runtime/merchant_trade_abi.h"
+#include "runtime/merchant/merchant_query_abi.h"
+#include "runtime/merchant/merchant_trade_abi.h"
 #include "runtime/player_sync_abi.h"
 #include "runtime/resource_domain.h"
 #include "runtime/terrain_abi.h"
